refactor: change project structure

+++ Frontend/api/Cart.ts
@@ -1,43 +1,7 @@
 "use client";
 import axios from "axios";
-import { CartItemDTO } from "./DTO/CartItemDTO";
-import { OrderDTO } from "./DTO/OrderDTO";
-import { UserDTO } from './DTO/UserDTO';
-import { axiosWithErrorHandling } from "./axiosConfig";
+import { axiosWithErrorHandling, Response } from "./axiosConfig";
 
-interface Product {
-    id: number;
-    title: string;
-    price: number;
-    qtyInStock: number;
-    weight: number;
-    image: string;
-    year: number;
-    category: string;
-    rushOrderEligible: boolean;
-}
-
-interface CartItem {
-    id: {
-        cartId: number;
-        productId: number;
-    };
-    cart: {
-        id: number;
-        subTotal: number;
-    };
-    product: Product;
-    qty: number;
-}
-
-interface CartResponse {
-    message: string;
-    cart: CartItem[];
-}
-interface Response {
-    data: any;
-    error: any;
-}
 const axiosInstance = axios.create({
     baseURL: `${process.env.NEXT_PUBLIC_API_URL}/cart`,
     timeout: 5000,
@@ -45,10 +9,15 @@ const axiosInstance = axios.create({
         "Content-Type": "application/json",
     },
 });
-
+const isValidCartId = (cartId: string | number | null): boolean => {
+    if (!cartId) return false;
+    if (typeof cartId === "number") return true;
+    const numberRegex = /^[0-9]+$/;
+    return typeof cartId === "string" && numberRegex.test(cartId);
+};
 
 export const getEmptyCart = async (): Promise<Response> => {
-    const { data, error } = await axiosWithErrorHandling(axiosInstance,{
+    const { data, error } = await axiosWithErrorHandling(axiosInstance, {
         method: "GET",
         url: `/new`,
     });
@@ -60,17 +29,19 @@ export const getEmptyCart = async (): Promise<Response> => {
 };
 export const getCartItems = async (): Promise<Response> => {
     const cartId = localStorage.getItem("cartId");
-    if (!cartId) {
+    if (!isValidCartId(cartId)) {
         await getEmptyCart();
         return {
             data: [],
             error: new Error("Cart is empty"),
         };
     }
-    return getCartItemsByCartId(cartId);
+    return getCartItemsByCartId(cartId as string);
 };
-export const getCartItemsByCartId = async (cartId: string | number): Promise<Response> => {
-    const { data, error } = await axiosWithErrorHandling(axiosInstance,{
+export const getCartItemsByCartId = async (
+    cartId: string | number
+): Promise<Response> => {
+    const { data, error } = await axiosWithErrorHandling(axiosInstance, {
         method: "GET",
         params: {
             cartId,
@@ -80,13 +51,13 @@ export const getCartItemsByCartId = async (cartId: string | number): Promise<Res
         data: error ? [] : data,
         error,
     };
-}
+};
 
 export const updateCartItemQty = async (
     productId: string,
     quantity: number
 ) => {
-    const { error } = await axiosWithErrorHandling(axiosInstance,{
+    const { error } = await axiosWithErrorHandling(axiosInstance, {
         method: "PUT",
         url: `/${localStorage.getItem("cartId")}`,
         data: {
@@ -98,17 +69,37 @@ export const updateCartItemQty = async (
 };
 
 export const deleteCartItem = async (productId: string) => {
-    return axiosWithErrorHandling(axiosInstance,{
+    if (!isValidCartId(localStorage.getItem("cartId"))) {
+        await getEmptyCart();
+        return {
+            error: new Error("Invalid request"),
+        };
+    }
+    const { data, error } = await axiosWithErrorHandling(axiosInstance, {
         method: "DELETE",
         url: `/${localStorage.getItem("cartId")}`,
         data: {
             productId,
         },
     });
+    if (
+        error &&
+        error.response.status === 404 &&
+        error.response.data.startsWith("Cart")
+    ) {
+        await getEmptyCart();
+        return {
+            error: new Error("Invalid request"),
+        };
+    }
+    return {
+        data,
+        error,
+    };
 };
 
 export const getTaxRate = async (): Promise<number> => {
-    const { data, error } = await axiosWithErrorHandling(axiosInstance,{
+    const { data, error } = await axiosWithErrorHandling(axiosInstance, {
         method: "GET",
         url: `/tax`,
     });
@@ -119,7 +110,10 @@ export const addItemToCart = async (
     productId: number,
     quantity: number
 ): Promise<Response> => {
-    const {data, error} =  await axiosWithErrorHandling(axiosInstance,{
+    if (!isValidCartId(localStorage.getItem("cartId"))) {
+        await getEmptyCart();
+    }
+    const { data, error } = await axiosWithErrorHandling(axiosInstance, {
         method: "POST",
         url: `/${localStorage.getItem("cartId")}/add`,
         data: {
@@ -127,7 +121,11 @@ export const addItemToCart = async (
             quantity,
         },
     });
-    if (error && error.response.status === 404 && error.response.data.startsWith("Cart")) {
+    if (
+        error &&
+        error.response.status === 404 &&
+        error.response.data.startsWith("Cart")
+    ) {
         await getEmptyCart();
         return addItemToCart(productId, quantity);
     }
@@ -135,7 +133,7 @@ export const addItemToCart = async (
         data,
         error,
     };
-}
+};
 // export const updateCart = async (productId: string, qty: number) => {
 //     try {
 //         await axios.put(`${apiBaseUrl}`, {
